\name{toOrdinal-package}
\alias{toOrdinal-package}
\docType{package}
\title{Function for converting cardinal to ordinal numbers for use with multiple languages
}
\description{
Function for converting cardinal to ordinal numbers for use with multiple languages
}

\details{
\tabular{ll}{
Package: \tab toOrdinal\cr
Type: \tab Package\cr
Version: \tab 0.0-1\cr
Date: \tab 2015-3-4\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr
}
}

\author{Damian W. Betebenner \email{DBetebenner@nciea.org}} 

\keyword{package}
