\name{toOrdinal-package}
\alias{toOrdinal-package}
\docType{package}
\title{Cardinal to Ordinal Number Conversion
}
\description{
Language specific cardinal to ordinal number conversion.
}

\details{
\tabular{ll}{
Package: \tab toOrdinal\cr
Type: \tab Package\cr
Version: \tab 1.3-1.0\cr
Date: \tab 2022-11-4\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr
}
}

\author{Damian W. Betebenner \email{DBetebenner@nciea.org}} 

\keyword{package}
