\name{toOrdinal-package}
\alias{toOrdinal-package}
\docType{package}
\title{Function for converting cardinal to ordinal numbers by adding a language specific ordinal indicator to the number
}
\description{
Function for Function for converting cardinal to ordinal numbers by adding a language specific ordinal indicator to the number. Function utilizes indicators as described in http://en.wikipedia.org/wiki/Ordinal_indicator.
}

\details{
\tabular{ll}{
Package: \tab toOrdinal\cr
Type: \tab Package\cr
Version: \tab 0.0-1\cr
Date: \tab 2015-3-10\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr
}
}

\author{Damian W. Betebenner \email{DBetebenner@nciea.org}} 

\keyword{package}
