\name{toOrdinal-package}
\alias{toOrdinal-package}
\docType{package}
\title{Cardinal to Ordinal Number Conversion
}
\description{
Function for converting cardinal to ordinal numbers by adding a language specific ordinal indicator (http://en.wikipedia.org/wiki/Ordinal_indicator) to the number.
}

\details{
\tabular{ll}{
Package: \tab toOrdinal\cr
Type: \tab Package\cr
Version: \tab 1.0-0.0\cr
Date: \tab 2017-6-25\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr
}
}

\author{Damian W. Betebenner \email{DBetebenner@nciea.org}} 

\keyword{package}
