\name{toOrdinal-package}
\alias{toOrdinal-package}
\docType{package}
\title{Function for Converting Cardinal to Ordinal Numbers by Adding a Language Specific Ordinal Indicator to the Number
}
\description{
Function for converting cardinal to ordinal numbers by adding a language specific ordinal indicator (http://en.wikipedia.org/wiki/Ordinal_indicator) to the number.
}

\details{
\tabular{ll}{
Package: \tab toOrdinal\cr
Type: \tab Package\cr
Version: \tab 0.0-1\cr
Date: \tab 2015-3-11\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr
}
}

\author{Damian W. Betebenner \email{DBetebenner@nciea.org}} 

\keyword{package}
